feat: support feedback mode for table
