+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/simple_table/simple_table_widgets/simple_table_reorder_button.dart
@@ -1,0 +1,77 @@
+import 'package:appflowy/generated/flowy_svgs.g.dart';
+import 'package:appflowy/plugins/document/presentation/editor_plugins/plugins.dart';
+import 'package:flutter/material.dart';
+
+class SimpleTableDraggableReorderButton extends StatelessWidget {
+  const SimpleTableDraggableReorderButton({
+    super.key,
+    required this.index,
+    required this.isShowingMenu,
+    required this.type,
+  });
+
+  final int index;
+  final ValueNotifier<bool> isShowingMenu;
+  final SimpleTableMoreActionType type;
+
+  @override
+  Widget build(BuildContext context) {
+    return Draggable<int>(
+      data: index,
+      feedback: Container(
+        color: Colors.red,
+        width: 100,
+        height: 100,
+      ),
+      child: GestureDetector(
+        behavior: HitTestBehavior.opaque,
+        onTap: () {},
+        child: SimpleTableReorderButton(
+          isShowingMenu: isShowingMenu,
+          type: type,
+        ),
+      ),
+    );
+  }
+}
+
+class SimpleTableReorderButton extends StatelessWidget {
+  const SimpleTableReorderButton({
+    super.key,
+    required this.isShowingMenu,
+    required this.type,
+  });
+
+  final ValueNotifier<bool> isShowingMenu;
+  final SimpleTableMoreActionType type;
+
+  @override
+  Widget build(BuildContext context) {
+    return ValueListenableBuilder(
+      valueListenable: isShowingMenu,
+      builder: (context, isShowingMenu, child) {
+        return MouseRegion(
+          cursor: SystemMouseCursors.click,
+          child: Container(
+            decoration: BoxDecoration(
+              color: isShowingMenu
+                  ? context.simpleTableMoreActionHoverColor
+                  : Theme.of(context).colorScheme.surface,
+              borderRadius: BorderRadius.circular(8.0),
+              border: Border.all(
+                color: context.simpleTableMoreActionBorderColor,
+              ),
+            ),
+            height: 16.0,
+            width: 16.0,
+            child: FlowySvg(
+              type.reorderIconSvg,
+              color: isShowingMenu ? Colors.white : null,
+              size: const Size.square(16.0),
+            ),
+          ),
+        );
+      },
+    );
+  }
+}
